post step: perform its own OIDC exchange for the envelope lookup instead of persisting the workflow token to runner state
